Aggiunta implementazione salvataggio file Utenti
Da testare le nuove funzionalità, in attesa di aggiunta gestione Utenti tramite menù

+++ src/hash.c
@@ -1,9 +1,11 @@
 #include "hash.h"
 
-unsigned int hash_djb2(const char *str) {
-    unsigned long hash = 5381;
+unsigned int hash_djb2(const char* str) {
+    unsigned int hash = 5381;
     int c;
+
     while ((c = *str++))
-        hash = ((hash << 5) + hash) + c; // hash * 33 + c
-    return (unsigned int)hash;
+        hash = ((hash << 5) + hash) + c;
+
+    return hash;
 }
